third_party/blink/web_tests/external/wpt/fetch/private-network-access/service-worker.https.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public to local: failed preflight. assert_equals: register error expected (string) "TypeError" but got (undefined) undefined
FAIL treat-as-public to local: success. assert_equals: register error expected (undefined) undefined but got (string) "TypeError"
FAIL treat-as-public to private: failed preflight. assert_equals: register error expected (string) "TypeError" but got (undefined) undefined
FAIL treat-as-public to private: success. assert_equals: register error expected (undefined) undefined but got (string) "TypeError"
PASS public to public: success.
Harness: the test ran to completion.
